+++ test/unit/scraper_test.rb
@@ -31,20 +31,44 @@ class ScraperTest < BaseTest
       expected = JSON.parse(load_fixture("expected-array.json"))
       assert_equal json["artworks"].size, expected["artworks"].size
 
-      pattern = {
-        artworks: [
-          {
-            name: String,
-            extensions: wildcard_matcher,
-            link: /^https:\/\/www\.google\.com\//,  # Matches the 'link' URL pattern
-            image: String #/https:\/\/encrypted-tbn\d\.gstatic\.com\/images\?q=tbn:/  # Matches the 'image' URL pattern
-          }
-        ] * 47
-      }
-
-      assert_json_match pattern, json
+      assert_json_match pattern(json), json
     end
 
   end
 
+  context "Klimt" do
+    setup do
+      @scraper = GoogleImageParser.new(load_fixture("klimt.html"))
+    end
+
+    should "output json" do
+      assert json = JSON.parse(@scraper.to_json)
+      assert_json_match pattern(json), json
+    end
+  end
+
+  context "Georgia O'Keeffe" do
+    setup do
+      @scraper = GoogleImageParser.new(load_fixture("okeeffe.html"))
+    end
+
+    should "output json" do
+      assert json = JSON.parse(@scraper.to_json)
+      assert_json_match pattern(json), json
+    end
+  end
+
+
+  def pattern(json)
+  {
+      artworks: [
+        {
+          name: String,
+          extensions: wildcard_matcher,
+          link: /^https:\/\/www\.google\.com\//,  # Matches the 'link' URL pattern
+          image: String #/https:\/\/encrypted-tbn\d\.gstatic\.com\/images\?q=tbn:/  # Matches the 'image' URL pattern
+        }
+      ] * json["artworks"].size
+    }
+  end
 end
